refactor: Fixed bugs and added github action (#2)
* Fixed bugs

* Fixed cr

* Added more comments

+++ tools/update_amazon.sh
@@ -1,70 +1,83 @@
-#!/bin/bash
+#!/bin/bash -x
 
-# Downloading mirror list file which holds the current address of the ftp server
-wget https://amazonlinux-2-repos-us-east-2.s3.dualstack.us-east-2.amazonaws.com/2/core/latest/debuginfo/x86_64/mirror.list
-# Downloading the latest compressed packages list
+# Extracting from the remote bucket the kernel versions we already handled (either successfully or failed).
+# Thus we will look for kernel we didn't handle it, so we will reduce run time of the script.
+# The line gets a list of all kernels in the remote bucket and extracts the kernel version from their name.
+gs_names=$(gsutil ls gs://btfhub/amzn/2/x86_64/ | sed 's,gs://btfhub/amzn/2/x86_64/,,g' | sed 's/.btf.tar.xz//g' | sed 's/.failed//g' | sort)
+
+repository=https://amazonlinux-2-repos-us-east-2.s3.dualstack.us-east-2.amazonaws.com/2/core/latest/debuginfo/x86_64/mirror.list
+
+# Downloading from a fixed location the URL for the remote repository holding the packages.
+echo "INFO: downloading ${repository} mirror list"
+wget $repository
+echo "INFO: downloading ${repository} information"
+# Downloading from the remote repository a compress sqlite3 file that contains a table with all kernel packages and their remote location.
 wget "$(head -1 mirror.list)/repodata/primary.sqlite.gz"
 rm -f mirror.list
-# Unzip the packages list
+
+# Extracting the compressed sqlite file.
 gzip -d primary.sqlite.gz
 rm -f primary.sqlite.gz
-# Extracting the kernel packages into a file
+
+# primary.sqlite contains a table with all packages can be downloaded. We look for packages with name that contains `kernel-debuginfo` but does not contain `common`.
+# We strip `..` from their location.
 packages=$(sqlite3 primary.sqlite "select location_href FROM packages WHERE name like 'kernel-debuginfo%' and name not like '%common%'" | sed 's#\.\./##g')
-# Creating a file with only the kernel versions
-echo $packages | tr "/" " " | awk '{print $3}' | sed 's/kernel-debuginfo-//g' | sed 's/.rpm//g' | sort > packages_names
-# Getting all kernel versions that we already produced them a BTF file
-gsutil ls gs://btfhub/amzn/2/x86_64/ | sed 's,gs://btfhub/amzn/2/x86_64/,,g' | sed 's/.btf.tar.xz//g' | sed 's/.failed//g' | sort > gs_names
-# Getting the new kernels that does not have a BTF file
-new_packages=$(comm -23 packages_names gs_names)
-rm -f gs_names packages_names primary.sqlite
-# For every new kernel
-for package in $new_packages; do
-    url=$(echo $packages | grep $package)
-    filename=$(basename "${url}")
+rm -f primary.sqlite
+
+# Iterating over the packages.
+for line in $packages; do
+    # Crafting the URL for downloading the debug symbols and the kernel version we are handling with.
+    url=${line}
+    filename=$(basename "${line}")
     # shellcheck disable=SC2001
     version=$(echo "${filename}" | sed 's:kernel-debuginfo-\(.*\).rpm:\1:g')
+
+    # Checking that we didn't handle that kernel version yet.
+    # If we did, we continue to the next kernel.
+    if [[ "${gs_names[*]}" =~ ${version} ]]; then
+        echo "INFO: file ${version}.btf already exists"
+        continue
+    fi
 
     echo URL: "${url}"
     echo FILENAME: "${filename}"
     echo VERSION: "${version}"
 
-    # Download kernel from FTP server
-    axel -4 -n 8 "http://amazonlinux.us-east-1.amazonaws.com/${url}"
-    mv "${filename}" "${version}.rpm"
+    # Parallel downloading of the kernel.
+    axel -4 -n 16 "http://amazonlinux.us-east-1.amazonaws.com/${url}" -o ${version}.rpm
     if [ ! -f "${version}.rpm" ]; then
-      echo "WARN: ${version}.rpm could not be downloaded"
-      continue
+        echo "WARN: ${version}.rpm could not be downloaded"
+        continue
     fi
 
-    # Extract the vmlinux location from the kernel image
+    # Extracting vmlinux file from rpm package.
     vmlinux=.$(rpmquery -qlp "${version}.rpm" 2>&1 | grep vmlinux)
     echo "INFO: extracting vmlinux from: ${version}.rpm"
-    # Extract the vmlinux into a file
     rpm2cpio "${version}.rpm" | cpio --to-stdout -i "${vmlinux}" > "./${version}.vmlinux" || \
     {
         echo "WARN: could not deal with ${version}, cleaning and moving on..."
+        rm -rf "./usr"
         rm -rf "${version}.rpm"
         rm -rf "${version}.vmlinux"
-        touch "${version}.failed"
         gsutil cp "${version}.failed" gs://btfhub/amzn/2/x86_64/"${version}.failed"
         continue
     }
 
-    # Generate BTF raw file from DWARF data
+    # Extracting the full BTF from the vmlinux file.
     echo "INFO: generating BTF file: ${version}.btf"
     pahole --btf_encode_detached "${version}.btf" "${version}.vmlinux"
-    # Compress the BTF
+    # Compressing the BTF.
     tar cvfJ "./${version}.btf.tar.xz" "${version}.btf"
+    # Uploading it to the bucket.
+    gsutil cp ./${version}.btf.tar.xz gs://btfhub/amzn/2/x86_64/${version}.btf.tar.xz
 
     rm "${version}.rpm"
     rm "${version}.btf"
+    rm "${version}.btf.tar.xz"
     rm "${version}.vmlinux"
-    # Move the compress BTF to the cloud
-    gsutil cp ./${version}.btf.tar.xz gs://btfhub/amzn/2/x86_64/${version}.btf.tar.xz
-    rm ${version}.btf.tar.xz
+
 done
 
-echo "Done"
 rm -f packages
 
 exit 0
